OMG it finally worked.
Stupid mistake making debug take in byte instead of int reported
erroneous zeroes when passing in int to debug.  Also has to pass
pointers instead of objects so values can be changed.

+++ ArduinoProjects/Suduino_04/Debug.h
@@ -1,18 +1,18 @@
 class Debug
 {
   public:
-    void DebugNum(char label[], byte num);
-    void DebugNum2(char label[], byte num1, byte num2);
+    void DebugNum(char label[], unsigned int num);
+    void DebugNum2(char label[], unsigned int num1, unsigned int num2);
   private:
 };
 
-void Debug::DebugNum(char label[], byte num)
+void Debug::DebugNum(char label[], unsigned int num)
 {
   Serial.print(label);
   Serial.println(num);
 }
 
-void Debug::DebugNum2(char label[], byte num1, byte num2)
+void Debug::DebugNum2(char label[], unsigned int num1, unsigned int num2)
 {
   Serial.print(label);
   Serial.print(num1);
